consolidate platform utils and restrict shell:open scope
Extract shared getOs() and shortenPath() into src/lib/platform.ts,
removing duplicates from useSettings, SettingsView, useGZDoom, and
App.vue. Restrict shell:open regex to local file paths only. Wire
up platform-specific innoextract install hints in SettingsView.
