+++ src-tauri/tauri.conf.json
@@ -26,7 +26,7 @@
   },
   "plugins": {
     "shell": {
-      "open": "^.+$"
+      "open": "^(/|[A-Za-z]:\\\\)"
     }
   },
   "bundle": {
